services/api-gateway/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
greenlet>=3.0.0
alembic>=1.13.1

# ── Cache ─────────────────────────────────────────────────────────────────────
redis[hiredis]>=5.0.4

# ── Auth / Security ───────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.2
passlib>=1.7.4

# ── HTTP client ───────────────────────────────────────────────────────────────
httpx>=0.27.0
tenacity>=8.3.0

# ── Validation ────────────────────────────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.2.0
email-validator>=2.1.1

# ── Logging ───────────────────────────────────────────────────────────────────
structlog>=24.1.0

# ── Rate limiting ─────────────────────────────────────────────────────────────
slowapi>=0.1.9
limits>=3.10.1

# ── Misc ─────────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
orjson>=3.10.0
